+++ spec/pb_kits/playbook/kits/textarea_spec.rb
@@ -21,10 +21,10 @@ RSpec.describe Playbook::PbTextarea::Textarea do
 
   describe "#classname" do
     it "returns namespaced class name", :aggregate_failures do
-      expect(subject.new({}).classname).to eq "pb_textarea_kit"
-      expect(subject.new({dark: true}).classname).to eq "pb_textarea_kit_dark"
-      expect(subject.new({error: "Something is wrong"}).classname).to eq "pb_textarea_kit error"
-      expect(subject.new({dark: true, error: "Something is wrong"}).classname).to eq "pb_textarea_kit_dark error"
+      expect(subject.new({}).classname).to eq "pb_textarea_kit resize_none"
+      expect(subject.new({dark: true}).classname).to eq "pb_textarea_kit_dark resize_none"
+      expect(subject.new({error: "Something is wrong"}).classname).to eq "pb_textarea_kit error resize_none"
+      expect(subject.new({dark: true, error: "Something is wrong"}).classname).to eq "pb_textarea_kit_dark error resize_none"
     end
   end