+++ app/pb_kits/playbook/pb_text_input/text_input.rb
@@ -31,7 +31,7 @@ module Playbook
       end
 
       def validation_pattern
-        validation[:pattern] || ""
+        validation[:pattern] || nil
       end
 
       def data